doc(dogfood): update comments in EndBlock
...and refactor getters / setters to `pending.go`.
